server/runtime/pyproject.toml
[project]
name = "huxley"
version = "0.1.0"
description = "Huxley — voice agent framework runtime."
requires-python = ">=3.13"
dependencies = [
    "huxley-sdk",
    # First-party skills are bundled here so `uv sync` installs them and
    # the framework can resolve their `huxley.skills` entry points. Core
    # never imports them statically — discovery is via entry points only.
    # Stage 4 will let a persona reference additional pip-installable
    # skills (`huxley-skill-<name>`); the persona's deps drive installation,
    # and these two stay listed only because they ship with the framework.
    "huxley-skill-audiobooks",
    "huxley-skill-system",
    "openai>=1.60",
    "websockets>=14.0",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "structlog>=24.4",
]

# openwakeword is installed separately on the Pi (tflite-runtime has no 3.13 wheels).

[project.scripts]
huxley = "huxley.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/huxley"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires external services (API keys, ffmpeg, etc.)",
]
